Add the frontend - view files.
index.html: The home page of the web application that displays people's cards.

insertion.html: A form to add a new person to the database in conjunction with the backend.

update.html: A form to update an existing person in conjunction with the backend.

Also some .css files are added to improve the user experience.